Refactor tomato/main.py into tomato/tomato/ module
