+++ tests/test_01_tomato.py
@@ -5,15 +5,7 @@ import subprocess
 import pytest
 import zmq
 
-from tomato.main import (
-    tomato_status,
-    tomato_start,
-    tomato_init,
-    tomato_reload,
-    tomato_pipeline_eject,
-    tomato_pipeline_load,
-    tomato_pipeline_ready,
-)
+from tomato import tomato
 from tomato.models import Reply
 
 PORT = 12345
@@ -21,14 +13,14 @@ CTXT = zmq.Context()
 
 
 def test_tomato_status_down():
-    ret = tomato_status(port=PORT, timeout=1000, context=CTXT)
+    ret = tomato.status(port=PORT, timeout=1000, context=CTXT)
     print(f"{ret=}")
     assert ret.success == False
     assert "tomato not running" in ret.msg
 
 
 def test_tomato_status_up(tomato_daemon):
-    ret = tomato_status(port=PORT, timeout=1000, context=CTXT)
+    ret = tomato.status(port=PORT, timeout=1000, context=CTXT)
     print(f"{ret=}")
     assert ret.success
     assert len(ret.data) == 2
@@ -36,7 +28,7 @@ def test_tomato_status_up(tomato_daemon):
 
 def test_tomato_start_no_init(datadir, stop_tomato_daemon):
     os.chdir(datadir)
-    ret = tomato_start(port=PORT, timeout=1000, context=CTXT, appdir=".")
+    ret = tomato.start(port=PORT, timeout=1000, context=CTXT, appdir=".")
     print(f"{ret=}")
     assert ret.success == False
     assert "settings file not found" in ret.msg
@@ -44,16 +36,16 @@ def test_tomato_start_no_init(datadir, stop_tomato_daemon):
 
 def test_tomato_start_with_init(datadir, stop_tomato_daemon):
     os.chdir(datadir)
-    ret = tomato_init(appdir=".", datadir=".")
+    ret = tomato.init(appdir=".", datadir=".")
     assert ret.success
-    ret = tomato_start(port=PORT, timeout=1000, context=CTXT, appdir=".")
+    ret = tomato.start(port=PORT, timeout=1000, context=CTXT, appdir=".")
     print(f"{ret=}")
     assert ret.success
 
 
 def test_tomato_start_double(datadir, stop_tomato_daemon):
     test_tomato_start_with_init(datadir, stop_tomato_daemon)
-    ret = tomato_start(port=PORT, timeout=1000, context=CTXT, appdir=".")
+    ret = tomato.start(port=PORT, timeout=1000, context=CTXT, appdir=".")
     print(f"{ret=}")
     assert ret.success == False
     assert f"port {PORT} is already in use" in ret.msg
@@ -65,10 +57,10 @@ def test_tomato_reload(datadir, stop_tomato_daemon):
         jsdata = json.load(inf)
     with open("devices.yml", "w") as ouf:
         yaml.dump(jsdata, ouf)
-    ret = tomato_status(port=PORT, timeout=1000, context=CTXT)
+    ret = tomato.status(port=PORT, timeout=1000, context=CTXT)
     assert ret.success
     assert len(ret.data) == 2
-    ret = tomato_reload(port=PORT, timeout=1000, context=CTXT, appdir=".")
+    ret = tomato.reload(port=PORT, timeout=1000, context=CTXT, appdir=".")
     print(f"{ret=}")
     assert ret.success
     assert len(ret.data) == 1
@@ -76,7 +68,7 @@ def test_tomato_reload(datadir, stop_tomato_daemon):
 
 def test_tomato_pipeline(datadir, stop_tomato_daemon):
     test_tomato_start_with_init(datadir, stop_tomato_daemon)
-    ret = tomato_pipeline_load(
+    ret = tomato.pipeline_load(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -89,7 +81,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
     assert ret.data.sampleid == "test"
     assert ret.data.ready == False
 
-    ret = tomato_pipeline_load(
+    ret = tomato.pipeline_load(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -102,7 +94,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
     assert "pipeline dummy-5 is not empty" in ret.msg
     assert ret.data.sampleid == "test"
 
-    ret = tomato_pipeline_ready(
+    ret = tomato.pipeline_ready(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -114,7 +106,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
     assert ret.data.sampleid == "test"
     assert ret.data.ready
 
-    ret = tomato_pipeline_ready(
+    ret = tomato.pipeline_ready(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -126,7 +118,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
     assert ret.data.sampleid == "test"
     assert ret.data.ready
 
-    ret = tomato_pipeline_eject(
+    ret = tomato.pipeline_eject(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -138,7 +130,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
     assert ret.data.sampleid is None
     assert ret.data.ready == False
 
-    ret = tomato_pipeline_eject(
+    ret = tomato.pipeline_eject(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -153,7 +145,7 @@ def test_tomato_pipeline(datadir, stop_tomato_daemon):
 
 def test_tomato_pipeline_invalid(datadir, stop_tomato_daemon):
     test_tomato_start_with_init(datadir, stop_tomato_daemon)
-    ret = tomato_pipeline_load(
+    ret = tomato.pipeline_load(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -165,7 +157,7 @@ def test_tomato_pipeline_invalid(datadir, stop_tomato_daemon):
     assert ret.success == False
     assert "pipeline bogus not found" in ret.msg
 
-    ret = tomato_pipeline_eject(
+    ret = tomato.pipeline_eject(
         port=PORT,
         timeout=1000,
         context=CTXT,
@@ -176,7 +168,7 @@ def test_tomato_pipeline_invalid(datadir, stop_tomato_daemon):
     assert ret.success == False
     assert "pipeline bogus not found" in ret.msg
 
-    ret = tomato_pipeline_ready(
+    ret = tomato.pipeline_ready(
         port=PORT,
         timeout=1000,
         context=CTXT,
